requirements.txt
astor==0.7.1
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Migrate==2.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SocketIO==3.0.2
Flask-SQLAlchemy==2.3.2
gast==0.2.0
gensim==3.6.0
gevent==1.3.7
graphviz==0.9
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jwcrypto==0.6.0
matplotlib==2.2.3
nltk==3.3
numpy==1.15.4
pandas==0.23.4
passlib==1.7.1
psycopg2==2.7.6.1
PyJWT==1.6.4
requests==2.20.1
scikit-learn==0.20.0
scipy==1.1.0
simplejson==3.13.2
SQLAlchemy==1.2.13
SpeechRecognition==3.8.1
google-api-python-client==1.7.4
google-cloud-speech==0.36.0